breaking out seperate screens for react navigator
