+++ android/settings.gradle
@@ -1,6 +1,5 @@
 rootProject.name = 'AlarmApp'
 
-
 include ':app'
 include ':react-native-maps'
 project(':react-native-maps').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-maps/lib/android')
