+++ package.json
@@ -20,8 +20,9 @@
   "dependencies": {
     "react": "16.0.0",
     "react-native": "0.51.0",
-    "react-native-geocoder": "^0.5.0",
+    "react-native-google-places-autocomplete": "^1.3.6",
     "react-native-maps": "0.20.0",
-    "react-native-video": "^2.0.0"
+    "react-native-video": "^2.0.0",
+    "react-navigation": "^1.5.11"
   }
 }
